+++ packages/ui/src/components/PageControl/components/PageControlProgress/PageControlProgress.tsx
@@ -1,8 +1,9 @@
-import { AnimationEventHandler, FC, AnimationEvent, Ref, useRef, useEffect, useState, CSSProperties } from 'react';
+import { AnimationEventHandler, FC, Ref, useRef, useEffect, useState, CSSProperties } from 'react';
 import style from './style.module.css'
 import { PageControlItem, PageControlItemProps } from '../PageControlItem';
-import { csx, mergeRefs } from '@vega-ui/utils';
+import { csx, mergeEventHandlers, mergeRefs } from '@vega-ui/utils';
 import { usePageControlContext } from '../../hooks';
+import { PageControlSize } from '../../types.ts';
 
 export interface PageControlProgressProps extends PageControlItemProps {
   /**
@@ -40,30 +41,30 @@ export interface PageControlProgressProps extends PageControlItemProps {
    * If not provided, active state is inferred from context.
    */
   current?: boolean
+  
+  /**
+   * Defines the size of the page control items.
+   */
+  size?: PageControlSize
 }
 
 /** PageControlProgress is a UI component used within PageControl to represent a navigation item with a visual progress indicator. It animates over time and triggers a callback when the progress completes, enabling automatic step transitions. */
 export const PageControlProgress: FC<PageControlProgressProps> = ({
   className,
   onProgressEnd,
-  onAnimationEnd: _onAnimationEnd,
   index,
   duration = 5000,
   current,
+  size,
   style: _style,
   ref,
   ...props
 }) => {
-  const { active } = usePageControlContext()
+  const { active, size: _size } = usePageControlContext()
   const progressRef = useRef<HTMLButtonElement>(null)
   const [progressValue, setProgressValue] = useState(0)
   
   const isCurrent = current ?? (active === index)
-  
-  const onAnimationEnd = (e: AnimationEvent<HTMLButtonElement>) => {
-    _onAnimationEnd?.(e)
-    onProgressEnd?.(e)
-  }
   
   useEffect(() => {
     const el = progressRef.current
@@ -98,7 +99,8 @@ export const PageControlProgress: FC<PageControlProgressProps> = ({
       aria-valuemax={isCurrent ? 100 : undefined}
       aria-valuenow={isCurrent ? progressValue : undefined}
       className={csx(style.pageControlProgress, className)}
-      onAnimationEnd={onAnimationEnd}
+      size={size ?? _size}
+      onAnimationEnd={mergeEventHandlers(props.onAnimationEnd, onProgressEnd)}
       style={{
         '--duration': `${duration}ms`,
         ..._style