+++ packages/ui/src/components/PageControl/providers/PageControlProvider/PageControlProvider.tsx
@@ -7,12 +7,16 @@ export type PageControlProvider = PageControlContextState
 export const PageControlProvider: FC<PropsWithChildren<PageControlProvider>> = ({
   active,
   variant,
+  size,
+  itemRef,
   children
 }) => {
   const value = useMemo(() => ({
     active,
     variant,
-  }), [active, variant])
+    size,
+    itemRef,
+  }), [active, variant, size, itemRef])
 
   return (
     <PageControlContext.Provider value={value}>